Removing some deprecated functions, embedding them into useLanguage.ts and simplifying them
