+++ src/language/sharedLanguage.ts
@@ -4,8 +4,8 @@ import { marked } from 'marked';
 import { mangle } from 'marked-mangle';
 import type { HTMLReactParserOptions } from 'html-react-parser';
 
-import type { IUseLanguage, ValidLanguageKey } from 'src/hooks/useLanguage';
-import type { IAltinnOrgs, IApplication, IDataSources, ILanguage, ITextResource } from 'src/types/shared';
+import type { IUseLanguage } from 'src/hooks/useLanguage';
+import type { IAltinnOrgs, IApplication, IDataSources, ITextResource } from 'src/types/shared';
 
 marked.use(mangle());
 
@@ -26,60 +26,6 @@ DOMPurify.addHook('afterSanitizeAttributes', (node) => {
     }
   }
 });
-
-/**
- * @deprecated Use lang() from useLanguage.ts instead
- * @see useLanguage
- */
-function getLanguageFromKey<T extends ValidLanguageKey | undefined>(key: T, language: ILanguage | null) {
-  if (!key || !language) {
-    return key;
-  }
-  const path = key.split('.');
-  const value = getNestedObject(language, path);
-  if (!value || typeof value === 'object') {
-    return key;
-  }
-  return value;
-}
-
-function getNestedObject(nestedObj: ILanguage, pathArr: string[]) {
-  return pathArr.reduce((obj, key) => (obj && obj[key] !== 'undefined' ? obj[key] : undefined), nestedObj);
-}
-
-export type LangParams = (string | undefined | number)[];
-
-// Example: {getParsedLanguageFromKey('marked.markdown', language, ['hei', 'sann'])}
-/**
- * @deprecated Use lang() from useLanguage.ts instead
- * @see useLanguage
- */
-export function getParsedLanguageFromKey(
-  key: ValidLanguageKey | undefined,
-  language: ILanguage | null,
-  params?: LangParams,
-  stringOutput?: false,
-): JSX.Element | null;
-export function getParsedLanguageFromKey(
-  key: ValidLanguageKey | undefined,
-  language: ILanguage | null,
-  params?: LangParams,
-  stringOutput?: true,
-): string;
-export function getParsedLanguageFromKey(
-  key: ValidLanguageKey | undefined,
-  language: ILanguage | null,
-  params?: LangParams,
-  stringOutput?: boolean,
-): any {
-  const name = getLanguageFromKey(key, language);
-  const paramParsed = params ? replaceParameters(name, params) : name;
-
-  if (stringOutput) {
-    return paramParsed;
-  }
-  return getParsedLanguageFromText(paramParsed);
-}
 
 export const getParsedLanguageFromText = (
   text: string | undefined,
@@ -123,7 +69,8 @@ const replaceRootTag = (domNode: any) => {
   }
 };
 
-const replaceParameters = (nameString: string | undefined, params: LangParams) => {
+export type LangParams = (string | undefined | number)[];
+export const replaceParameters = (nameString: string | undefined, params: LangParams) => {
   if (nameString === undefined) {
     return nameString;
   }
@@ -135,28 +82,6 @@ const replaceParameters = (nameString: string | undefined, params: LangParams) =
   });
   return mutatingString;
 };
-
-/**
- * @deprecated Use lang() from useLanguage.ts instead
- * @see useLanguage
- */
-export function getTextResourceByKey<T extends string | undefined>(
-  key: T,
-  textResources: ITextResource[] | null,
-): string | T {
-  if (!textResources || !key) {
-    return key;
-  }
-
-  const textResource = textResources.find((resource: ITextResource) => resource.id === key);
-  if (!textResource) {
-    return key;
-  }
-  // Checks if this text resource is a reference to another text resource.
-  // This is a common case when using likert component
-  const resource = textResources.find((resource) => resource.id === textResource.value) || textResource;
-  return resource.value;
-}
 
 /**
  * Replaces all variables in text resources with values from relevant source.
